[CHORE] Move sw check logic into App
